Extend Keyboard LED support to 16 bits
 1 LED_NUM_LOCK
 2 LED_CAPS_LOCK
 3 LED_SCROLL_LOCK
 4 LED_COMPOSE
 5 LED_KANA
 6 LED_POWER
 7 LED_SHIFT
 8 LED_DO_NOT_DISTURB
 9 LED_MUTE
10 LED_TONE_ENABLE
11 LED_HIGH_CUT_FILTER
12 LED_LOW_CUT_FILTER
13 LED_EQ_ENABLE
14 LED_SOUND_FIELD_ON
15 LED_SURROUND_ON
16 LED_REPEAT

+++ src/SingleReport/BootKeyboard.h
@@ -34,7 +34,7 @@ class BootKeyboard_ : public PluggableUSBModule, public DefaultKeyboardAPI
 {
 public:
     BootKeyboard_(void);
-    uint8_t getLeds(void);
+    uint16_t getLeds(void);
     uint8_t getProtocol(void);
     void wakeupHost(void);
     
@@ -75,7 +75,7 @@ protected:
     uint8_t protocol;
     uint8_t idle;
     
-    uint8_t leds;
+    uint16_t leds;
     
     uint8_t* featureReport;
     int featureLength;
